Biggest change: MatsTrace hidden from user-facing API, burried further down into impl.
There has never been a good reason to delve into the actual MatsTrace,
other than debugging. It has thus been moved away from the "mats-lib"
project. I first left a husk of it as an interface, with the "Impl"
residing in the util-project. However, it went further..

When trying to decide how to move this, I realized that the dream of
some day doing this with binary serialization would be hard to do if
everything was assumed to be Strings.

Now, the logic is as follows: The serialization is utterly
implementation dependent. Whether it uses JSON, XML, or Google
Protobufs, is up the the implementation: The API only specifies what the
user expects his DTOs and STOs to be - and the entire protocol, and
whether to use the MatsTrace logic is up to the implementation.

The concept of MatsTrace has thus totally gone from the "specification"
part. It is now just a utility for implementations.

I've made two interfaces: MatsTrace and MatsSerializer. The Serializer
assumes byte-array for MatsTrace serialization, and a generified type
for the "Object serialization", i.e. STOs and DTOs.

I've made an implementation of these, employing Jackson for the
MatsTrace, but serializing directly to byte-array. For the objects
(STOs, DTOs), String-serialization with JSON is used. The rationale for
not using byte arrays here too is that since JSON is used on the
envelope, carrying bytes inside there is pretty horrendous, so it made
no sense. So the "inner" serialization of STOs and DTOs is plain JSON-
Strings, while the "outer"/"envelope" MatsTrace serialization is JSON-
but-directly-to-UTF-8.

These four interfaces/classes reside in the new 'mats-impl-serial'
project (although it should really have been divided into two, i.e. the
API-part and the (possibly) different impl-parts).

The JMS implementation only depends on the API-aspect of this, i.e. the
MatsTrace and MatsSerializer, not the implementations. The only place
the MatsSerializer_DefaultJson is used in "userland" is in the
Rule_Mats. It will also have to be used in actual user code, to
configure up the JmsMatsFactory, to configure up the MatsFactory.

The getTraceId() was moved from MatsTrace to
ProcessContext.getTraceId(), and the same for getFrom(), which now is
ProcessContext.getFromStageId(). Also, to get a "debug MatsTrace"
prettyprint, you just run toString() on the ProcessContext.

Man, this is much cleaner.