Added "newInstance" method on the MatsSerializer, since this concept is tied somewhat to the serialized in use.
This is needed when you start processing for an endpoint, as you need to
instantiate a new state object. GSON is more lenient with what it can
instantiate than Jackson.

The default impl in Mats JMS impl used to be a "find constructor, invoke
it". This is now moved out to the MatsSerializer.

Also cleans deps of mats-serial and mats-serial-json, since they don't
really need mats-lib.