Made the compress/uncompress methods non-static and protected, to allow for override.
It should be fairly simple to make an object pool out of these, by
employing a simple lock-free stack: When the stack is empty (as it is
upon start), make a new instance. Enpool it when done. This should
quickly lead to a steady-state of correct number of instances. Only
question is wether the CAS'es in the lock-free impl are heavier than the
GC that will otherwise have to pick up both the Deflaters (with their
bad finalize() method), and the byte arrays.