Bunch of changes: Most files because getInitiator() -> createInitiator(). Added waitForStarted()
Also:

Split "ProcessContext" into "ProcessDetachedContext" with extension
"ProcessContext" - so that the get-aspects of the context can be shared
out of the process lambda.

Added ability to set MatsDestinationPrefix.

Added MatsTrace.getTracePropertyKeys()

Added copying of TraceProperties when initiation from within a process
lambda.

Added mentioned StartStoppable.waitForStarted(), along with a test for
this.

Some copyedit and added javadocs.