Closes in on issue #22 - Session pooling. Continues issue #6 - priorities. Began issue #27: "VERY BAD!" scenarios.
Made the JmsMatsJmsSessionHandler_Pooling. Not finished.

Starting clean-up of Exceptions. Should not have hierarchy of
Exceptions. The MatsBackendException is gone, replaced by an JMS MATS
impl-specific JmsMatsJmsException instead. The
MatsRefuseMessageException is moved into the MatsEndpoint class.

Both this, and the issue of transaction demarcation in initiations, led
to a refactoring of how messages are sent: They are now first produced
and "enlisted" (as JmsMatsMessage instances), and then afterwards they
are sent. This enables an ability to try to re-send them if the
initiation runs into the "VERY BAD!"-scenario. This resulted in a nice
refactoring whereby the processing of the user-supplied Stage lambda
does not interact with the JMS API at all - only internal structures:
First needed stuff is picked off of the incoming MapMessage - then the
user lambda is run, where any new sendings are put on a "messagesToSend"
list. Then, after returning into the JMS MATS impl, this is then sent in
one go using the JMS API.