Closes #57, closes #58, closes #59, closes #62 plus a bugfix in Session pooler
57: Ability to name MatsFactory, and have SpringConfig set this to
beanName

58: Ability to get hold of all configured Endpoints and Initiators of a
MatsFactory

59: SpringConfig: Qualify which MatsFactory to use.

Downgrade Spring dependency to 4.3.+, just to ensure that we can handle
an older Spring. (It also handles 5.0 and 5.1)

Refactored @EnableMats with more distinct steps (along with the handling
of qualification).

Refactored MatsTestActiveMq to a instantiatable class, to also be able
to easier use it in SpringConfig test harness.

The bugfix was a situation where the sync-order verififaction kicked in.
Rearranged code so that creation of JMS Session is not done within sync.