Spring testing story. Closes #71, closes #12
For Issue #12 "Spring annotation for creating a MatsFactory", I've
currently decided that it is better to have a simple way to create a
Spring @Bean. The rationale for this is that the JmsMatsFactory actually
needs a JMS ConnectionFactory - and it is this that needs to be changed
between "regular" connectivity (to a central MQ), localhost connectivity
(to a localhost-running MQ, typically in development scenarios), and an
in-vm LocalVm MQ (dev, but particular integration testing). So I close
this for now.