Closing in on #74, @MatsClassMapping. Dang, that was close, was it not for a massive thinko..
Thinko: The state object is not serialized after the Stage lambda exits,
but when message-generating methods on the ProcessContext is invoked,
e.g. request/reply/initiate. Need to wrap the ProcessContext to null out
the "dependency-injected" fields upon entry of request, and re-inject
them after exit when going back into the lambda.