Closes #100: "MatsRefuseMessageException does not work". Also made shutdown a tad smoother.
The MatsRefuseMessageException now works, with tests.

In JmsMatsStageProcessor, the shutdown-by-closing-session is made a bit
more robust to handle situations arising in testing, where right after
the test, the system is shut down before the StageProcessor has gotten
into consumer.receive() yet, thus going into "phase 1" of the stop
procedure where it waits up to 30 seconds. It will therefore wait the
full 30 seconds, since the StageProcessor just entered the receive() and
is NOT "out doing work". This is the known async-site of the shutdown
procedure, but now a common situation (testing) should be handled a tad
better.

Also, some handling of ActiveMQs annoyances wrt. when the
consumer.receive() call is interrupted (as happens in "phase 2" of
shutdown).

Also, if consumer.receive() returns null and the run-flag actually is
false, we just break out right away, without closing, giving better
logging and eliding one unnecessary attempt to close the session.