Starts #106: WebSocket "bridge" to Mats: "MatsSockets"
Come quite some way. Pretty nice, if I must say so myself.

Implemented using the JSR-356 WebSocket API. A concept of
MatsSocketEndpoint is established. The is registered on the
MatsSocketServer. It is a Mats-variant of a WebSocket "endpoint" living
in a separate namespace to the MatsEndpoints, but which assumedly
typically will bridge to the MatsEndpoint of the same Id. For a
MatsSocketEndpoint, there is an "incomingForwarder" and "replyAdapter"
lambda registered. The forwarder shall check authorization and then
forward to the MatsEndpoint. When the MatsEndpoint replies, the
replyAdapter will map over to the expected DTO for the
MatsSocketEndpoint.