Working on #106: MatsSockets. Started the ClusterStoreAndForward functionality.
When doing a request on Mats, the reply might come to a different node
that originated it. We want reliable messaging, and can thus not use
replyToSubscription.

What we need to do, is to store the message in a "holding area" (default
impl will be a SQL db), and then check which node actually currently
holds the WebSocket - if any. We send a topic message to that node,
notifying it that there are messages. It will get it, retrieve the
messages, forward them over the WebSocket, and mark them as done (i.e.
deleting them) from the holding area.

Implemented now: Interface of ClusterStoreAndForward, "SingleNode"
variant of it implemented, and the MatsSocketServer code calls it -
everything still works, now with the added topic-jump.