Working on #106: MatsSockets. Promises. Refactored MatsSocketServer into several classes.
Also, now all SEND and RECEIVE will get an acknowledgement of whether
received. Introduced messageSequenceId for this, the idea is a TCP
window of outstanding messages. Since this is sequential, and all
processing is sequential, such an id is not a problem. Thus dropping
correlationId as a system-level idea: Now both
"outstandingSendOrRequest" is using this sequence, and also for promise
resolving. CorrelationId is only used for the app. It should be possible
to hammer in some resend-logic using this stuff, since we keep the
outstanding messages on client until acknowledged. The server can pick
up if he is missing a message, and ask for retransmit. The server can
also know what the latest sequence he has gotten (continuous sequence),
and thus if he gets an older (earlier) message than this, he can just
drop it on the floor, thus the client can just retransmit everything he
has outstanding on reconnect. This however implies that the server needs
to store this "latest handled sequence number".