TECHNICAL: Fixed build issue with javascript tests.
The JavaScript tests starts a server based on the code in the test
folder of this project. For this to run, we need to compile all
dependent projects, and this projects test classes. This so that
all the jars and classes we depen on are compiled. The previous commit
ommitted the compilation of the test classes.

Also added a commit-msg hook in CONTRIBUTE that can be used to
automatically sign all commits. This is based on the Gerrit commit-msg
hook, but instead adds a fixed string in accordance with the SCA.

I contribute this material in accordance with the signed SCA.