Refactoring Javascript build
Changed the javascript build script, so that it only deals with
Javascript, and not the start/stop of the server. This should make it
easier to integrate other clients into the integration test step.

The tests where also renamed in the gradle build file to be called
javascriptTest, and now calls a new testAll npm task. This task also
executes unit tests now.

There was also an accidental rename of the MatsSocket.js library name
that was fixed.

I contribute this material in accordance with the signed SCA.