Added Dart client for MatsSocket
Created a MatsSocket client in Dart, based on the Javascript library.
The code is slightly different, as it relies more on constructs found
in Dart, most notably Stream, Completer and Future, and the async
support in Dart.

This client fullfills the same unit and integration tests that the
Javascript library has. It has a mock version of the WebSocketChannel
for use in unit tests that allows easy development of a mock server.

I contribute this material in accordance with the signed SCA.