MatsSocket: Sequential-test: Had forgot .flush() -> Down to 2 ms per roundtrip.
When I changed to "autopipeline", there is a 2ms wait for new messages -
with an optional matsSocket.flush() to get it sent right away. I forgot
to add this in the continuous_sequential_requests. With this added,
we're back down to ~2ms per request-reply ("matsping", i.e. all the way
into the matsfabric and back). Nice.