Added support for mutable wsUrls in MatsSocket.
In order for clients to be reconfigurable, we need to expose wsUrls as a
mutable property, so that users of the library can create ui to change
the backend url.

Also changed the authorization logic to be contained within it's own
class.

I contribute this material in accordance with the signed SCA.