Synchronized the Dart library with JavaScript.
The Dart library should now be more in sync with the JavaScript library.
When we are connecting, we do an infinite incremental back-off to
connect to the WebSocket url. There is a difference here in Dart, as we
get the connection immediatly, and we receive the error on connection in
the stream of events, rather than in the connection future. To handle
this, we don't consider us connected until we receive the welcome
message back from the server.

Added EnvelopeSubType as an enum. This caused a bit of an issue with the
extension methods on EnvelopeType, as Dart was not able to actually
distinguish the 2 enum types as separate types, so some code changes
where done for parsing and formatting enums.

The wsUrls field was also made mutable via a setter. When the wsUrls are
set, we effectivly close and reset the connection.

The close code was changed to be aligned to the JavaScript close code,
actually copying much of the code and converting to Dart.

I contribute this material in accordance with the signed SCA.