TravisCI still sometimes bail on the "Test_TimeToLive.checkInfrastructure"..
Trying to add "nonPersistent(TTL=forever)" to the "flusing" message,
hoping that this will resolve it.

Note: I have NEVER experienced this on neither IntelliJ run, nor command
line. Might be TravisCI that jitters on the run container, but it is
still very strange that it is the "checkTestInfrastructure" that fails.
Now the checkTestInfrastructure should send 5 /identical/ messages, and
the consumer should most definitely have concurrency=1, so process one
by one message. The flushing message should then definitely come after
the four "test messages" (which in checkTestInfrastructure also have
TTL=forever).