MatsSocket: Working on #144, reliable messaging
Started on reliable messaging.

Three-way message passing:
Each side has an inbox and an outbox. The goal is to get an
"information-bearing message" from one side's outbox to the other side's
inbox. The outbox is to enable retransmission, and the inbox is to
hinder double deliveries.

To do this reliably, we have three step:
side A: Sends the message from outbox. Receiver puts it in his inbox.
side B: Sends RECEIVED, including the msgId. When A receives it, he
deletes his outbox entry
side A: Sends ACKACK, including the msgId. When B receives it, he
deletes his entry from the inbox.

Step 2 and 3 are not really time critical, in particular step 3 can be
done whenever there is capacity on the line.