MatsSocket: Better handling of double delivery Client-to-Server
It was somewhat bad in that it relied on the Mats rollback if it failed,
and otherwise it relied HEAVILY on idempotence of the incomingHandler to
do exactly the same as last time. This was cool enough, only that Mats
emits a massive Exception when this happens - on ERROR.

Now it will actually store away the last reply (unless ACK, which is
considered default, thus eliding extre DB update for hopefully
common-path). If we next time get unique constraint violation upon
storing id in inbox DB, we read up this stored message. If there is an
envelope there, it is answered, and we do not go into handleIncoming. If
it is not there, it was an ACK, so we answer that.

* added a _sessionClosed to guard against a situation where we async
could start reconnecting even after the MatsSocket was closed.