Removing special case for String as Input type, and String as Reply type
If you Replied with a String, it would be directly added, not JSON
serialized.

If you accepted a String, you would get whatever was there, meaning that
if it was a JSON serialized object, you would get the "raw JSON".

This was intended as a trick to "accept Object", in that you could then
deserialize it as whatever you wanted. However, that would totally screw
over changing the serialization mechanism to something like Protobuf.

With the advent of MatsObject, which solves the "accepting Object" in a
much smoother way, this feature is just the garbage it always was.

Since I believe no-one in the world is using this right functionality
right now (the users of Mats using their own special serializer), I just
remove it without shedding a tear.