Decided to move MatsSocketEnvelopeDto into API: MatsSocketServer.
The rationale for this was that I have wondered since day whether this
would be smart. I had already moved MessageType down there, to kinda
"document" the Wire Protocol, the only thing lacking was the actual DTO.
But I hated the Jackson-annotations. However, after gaining the
knowledge of Mixins, I have now moved it without trace of Jackson, and
instead mix that in in the implementation.

In addition, it became screamingly obvious that the MessageEvent and
LiveMatsSocketSession.getLastMessageEvent() logic was basically just
reimplementing this exact same DTO. So why not instead provide the
*exact* messages that went over the wire, instead of a "shadow" of it?