Closes #169 - Dart client update.
This brings the Dart client fully in sync with the JavaScript version.

The Dart client now supports both the DartVM backend, as well as the
Dart HTML backend for dart2js. This is also verified through all the
tests that will work both from DartVM and dart2js. As the Browser does
not have an environment, the backend URLs for those tests are fixed at
localhost:8080 and localhost:8081.

Because the dart2js tests require an installed Chrome browser, these
tests are not run as part of the build. These can be run via gradle by
setting the project platform for Dart:

./gradlew dartTest -Pplatform=chrome

Please note however, that the url for the backend is hardcoded, but this
should match what the dartTest starts anyways.

All the code that abstract away the native parts that Dart does not
already abstract away, has been placed in MatsSocketPlatform. This
serves as a bridge to either dart:io or dart:html depending on what is
available.

I contribute this material in accordance with the signed SCA.