Improve unit test for reauth
Changed the reauth test to assert the order in which things occur,
checking that we first get the ack, then the reauth, before finally
receiving the reply from the server.

Also changed all Completers internally in Mats to be sync Completers,
that way any listeners are triggered inline, rather than waiting for the
current thread to complete. This was earlier necessary at least for some
unit tests that where very strict on the ordering of events. By making
all Completers sync, all event ordering is deterministic.

Finally, changed the user agent to no longer be a future, as there was
no need for it to be, and having it as a future will complicate how we
set the clv part when authenticating.

I contribute this material in accordance with the signed SCA.