Minor cleanup, some MDCs, hopefully correct efficienty wrt. clearing out outbox
Previously, we always relied on a clear-out round giving 0 messages.
This is made a bit more efficient with checking whether we got less than
MAX number of messages from CSAF, thus indicating that we got'em all.
The logic with notification count increasing while we are doing a round
should still hold without that last mandatory get-returns-zero-messages.
This was just too ugly, as typically only 1 message will be forwarded,
but we always had to do TWO selects to make sure we had it cleared.