Eagerly remove running handler count before reauth.
In some rare cases, the new AUTH would come in before the currently
running handler fully completes, and removes itself from the count of
running notifications. To fix this, we need to eagerly clear the
currently running notification count before we send the reauth message.

I contribute this material in accordance with the signed SCA.