Create new lists rather than clear acks/nacks.
When sending acks and nacks, we where clearing the lists, rather than
creating new ones. This will cause an error downstream, as the same
instance is used for sending. By the time we are sending the acks over
the WebSocket, the list is already cleared, and we end up sending an
empty list.

I contribute this material in accordance with the signed SCA.