Fix nested initiations to use the same MatsInitiator
If we get the default initiator, and within that execution again call
getDefaultInitiator, this would result in a new separate MatsInitator,
that would commit independently of the previous initiator, which can
cause issues when the messages sent will need data commited from the
outer initiator. To fix this, we always set the
__stageDemarcatedMatsInitiate when we create an initiator, so that
nested calls will use the same initiate.

I contribute this material in accordance with the signed SCA.