Fixes #187: "too many consecutive connection attempts: Trying to create WebSocket"
This wasn't as smart as I had envisioned, and after considering it, I
effectively removed the feature - but made it into a "connectionAttempt"
counter instead, now available as a property on the ConnectionEvent
class. However, I let the "bail out" functionality remain whereby if it
has attempted for a full week, it stops - one can envision an absurd
situation whereby a "forgotten node server" just keeps on trying to
connect to a MatsSocket service long gone.