Closes #205: Remove the multiple-connection-getting-guard exception from wrapped DataSource
In the MonitorConnectionGettingDataSourceProxy in
JmsMatsTransactionManager_JmsAndSpringManagedSqlTx, upon
getConnection(), I had a check for whether you had gotten a Connection
earlier within the same transactional demarcation.

The idea was to catch wrong usage of the Spring transactional
demarcation, but with an understanding that there could be legitimate
reasons for this, where a user would want to actually have an
independent Connection to get out of the transactional demarcation of
the Mats stage. This worry turned out to be correct, and I had failed to
realize that it is fairly easy to specify such a situation within the
Spring tx system: Specify
TransactionDefinition#PROPAGATION_REQUIRES_NEW, which AFAIU ends up
fetching a new/separate Connection (albeit the JavaDoc talks about
"suspension", it seems like this is implemented by just performing the
new transaction within a separate Connection).

This turned out to be used several places in my main user's systems.