Closes #217: Remove IncomingHandler's explicit "defaultForward" methods, change semantics.
New methods:
* forwardEssential(toMatsEndpointId, msg)
* forwardNonessential(...)
* forward(toMatsEndpointId, msg, customInit)

.. which replaces:
* forwardInteractiveUnreliable(..)
* forwardInteractiveReliable(..)
* forwardCustom(..)
.. all of which did not have the "toMatsEndpointId" argument.

The rationale for this is that one should be explicit about which Mats
Endpoint one sends to, and not by default encourage to set up Mats
endpoints which are called exactly the same. The reason for this again
is that the namespaces for MatsSocket and Mats is totally different: On
a given "Mats fabric", there could potentially be many different
MatsSocketFactories - and each of those have their own namespace (i.e.
the MatsSocket EndpointIds are actually totally disjunct from the Mats
EndpointIds).

Along with this, I also changed the methods on MatsFuturizer, as those
were synchronized with the MatsSocket ones.