Fixes #227, Use specific MatsInitiator for MatsFuturizer
Currently, the MatsFuturizer uses matsFactory.getDefaultInitiator() to
get an initiator. This method has a certain "magicness" with it when
used within a Stage. This posed a problem for a particular usage
(employing MatsFuturizer within a stage, which should never be done).
But this got me thinking whether it would at any rate be better if a
MatsFuturizer utilized its own MatsInitiator.

Thus, change it to matsFuturizer.getOrCreateInitiator("MatsFuturizer" +
endpointIdPrefix).