Fixes #235: JmsSessionHolder is not recycled if exception raised inside initiate
Moving .release invocation to finally block, and handling in the
pooling-implementation (JmsMatsJmsSessionHandler_Pooling) it can be
invoked in such a finally block by ignoring the call if already closed
or crashed. This was already handled correctly in the
JmsMatsJmsSessionHandler_Simple. Also pointing out that .release() may
be used in such a finally-style handling in the JavaDoc.