Fix issue with authorization expiring during reconnect.
If we end up taking long enough time during the reconnect for the
authorization to expire, we will be stuck in a loop where we will never
be able to authenticate, as requests will likely fail due to the
authorization having expired.

To fix this, we eagerly check the authorization expiration on each
reconnect attempt, and request a new authorization token if it ends up
expiring.

I contribute this material in accordance with the signed SCA.