Closes #240, Handle if a SQL Connection from DataSourceUtils is "multi-wrapped", so that it is not our DeferredConnectionProxy
In a certain setup of Mats+Hibernate, it turned out that the SQL
Connection that was gotten via
DataSourceUtils.getConnection(_matsEmployedDataSource) was NOT of the
DeferredConnectionProxy type (i.e. our own), but instead was wrapped by
some Hibernate stuff.
This was obviously not handled well, since the Mats transaction manager
just casts it directly.
Implement using connection.isWrapperFor(DeferredConnectionProxy) with
subsequent connection.unwrap(DeferredConnectionProxy).