Closes #46, "stack overflow" and "call overflow" protection + completing logic with subinitiations
"overflow protection": If > 100 calls (including sub initiations), or
the stack height becomes > 25, break by (effectively) doing a
MatsRefuseMessageException, thus (insta-)DLQing.

completing sub initiation logic:

If you do a DefaultInitiator sub-initiation, it will be "hoisted" into
the existing transaction context.

If you do a non-Default Initiator sub-initiation, a new context will be
created (actually, an actual new Thread), and the initiation will go
there. However, a WithinStageContext will be transferred over to the new
Thread, so that the "total call number" will continusouly increase even
in such new context situations. This also makes it so that "from" and
"traceId" will be added to the new initiation, even though in a new
transactional context.

This logic "recurses".