Backing out of FactoryConfig.setName(..) return type change.
It used to return void, and I changed it to FactoryConfig to allow for
chaining. However, this is problematic to do in a revision, since the
JVM bytecode includes the return type as signature for a called method.
Thus, existing *compiled* code that invokes setName(..) will end up with
MethodNotFound exception when trying to invoke it if upgrading the
library, and not recompiling the calling code (e.g. when used with a
library that invokes setName for you).