Closes #6, priorities. Added n/2 StageProcessors that "SQL Selects" JMSPriority = 9
While developing the LocalHtmlInspectForMatsFactory with
LocalStatsMatsInterceptor, it became obvious that the "interactive"
functionality currently present (only setting JMSPriority to 9) didn't
quite make any dent: After sending 1000 messages to a multi-stage
service, and then throwing an "interactive" flow onto the same queue,
the latency of the latter quite obviously indicates that it is processed
after the first 1000.

Thus, now adding floor(n/2) stage processors where the selector is set
to "JMSPriority = 9". This seems to really rush those messages through.
The one annoying thing, is that many endpoints will never employ those
threads. But that should not really be that big a deal.

(Had to fix the Test_TimeToLive, which wrongly had the interactive flag
set - which thus suddenly got to employ the extra stage processor, not
only the 1 normal stage processor the test had specified (concurreny=1))